+++ services/reactionsTypeService.js
@@ -1,4 +1,6 @@
 const Repository = require('../repositories/reactionsTypeRepository');
+const ApiError = require("../utils/ApiError");
+const httpStatus = require("../utils/statusCodes");
 
 class ReactionsTypeService {
     async createReactionType(description) {
@@ -8,6 +10,8 @@ class ReactionsTypeService {
         return Repository.getAll();
     };
     async deleteReactionType(id) {
+        const reactionsType = await Repository.getById(id);
+        if (!reactionsType) throw new ApiError(httpStatus.NOT_FOUND, 'Reaction type not found');
         return Repository.delete(id);
     };
 }
